radically changed player movement concept (again)
The fully dynamic movement of the player (using tilted plates only)
was really hard to control. Instead a partially kinematic approach
is now used to control the player directly. The kinematic model is
simply a gyro with all rotation locked around the upwards axis.
However, the values for rotation and angular velocity are not directly
overridden by the kinematic model but mixed with the actual dynamic
values by a certain ratio (0=fully kinematic, 1=fully dynamic).
Even when this ratio is up to 0.99, the gyro is just stable enough to
be controlled. To simulate the decay of the gyro's "life" (or Energy),
the factor is calculated in a way that it increases when the gyro's
rotation slows down. However, the factor is never decreased, even if
the angular velocity should increase. When the factor exceeds 0.99
the controls are disabled and the values for rotation and angular
velocity are no longer manipulated and (gyro is dead x.x)